fix the bug in BlastFinder (crash on empty input)
makeblastdb doesn't like empty input file. It exits with
code 1 in that case. That is why it BlastRunner should be
omitted if the input set of sequences (consensus sequences)
is empty. There was an error in this logic: wrong blockset
was checked for emptyness.

This error was discovered on Windows on Yersinia genomes set.

[2015-Jun-13 21:00:37 UTC+03:00] FilterSeqs begin
[2015-Jun-13 21:00:37 UTC+03:00] FilterSeqs end
[2015-Jun-13 21:00:37 UTC+03:00] BlastFinder begin
[2015-Jun-13 21:00:37 UTC+03:00] RawWrite begin
[2015-Jun-13 21:00:38 UTC+03:00] RawWrite end
[2015-Jun-13 21:00:38 UTC+03:00] BlastRunner begin
std::exception: 'C:\\Yersinia_pestis_npge_32\\makeblastdb.exe
failed with code 1. Command: C:\\Yersinia_pestis_npge_32\\makeblastdb.exe
-dbtype nucl -out C:\\Users\\836D~1\\AppData\\Local\\Temp\\2\\npge_aa56ddb468
-in C:\\Users\\836D~1\\AppData\\Local\\Temp\\2\\npge_ac45360f05
-logfile /dev/null'