add CMake variable NPGE_LUA_CMD
Previously Lua command was deduced from Lua library
implementation used (PUC-RIO Lua => lua, LuaJIT => luajit).
This approach doesn't work if library exists and command
doesn't exist or of it's name is not "lua" or "luajit".

Default value for NPGE_LUA_CMD is "lua". To change it,
add the following option to cmake:

    -DNPGE_LUA_CMD=luajit-2.0.3